+++ CHANGELOG.md
@@ -1,2 +1,3 @@
 # v2.7.38
 
+* added context manager to SelfClosingGenerator